% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcurseR-package.R
\docType{package}
\name{calcurseR-package}
\alias{calcurseR}
\alias{calcurseR-package}
\title{calcurseR: Functions to use 'calcurse' as a general organising tool}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mpadge/calcurseR}
  \item Report bugs at \url{https://github.com/mpadge/calcurseR/issues}
}

}
\author{
\strong{Maintainer}: Mark Padgham \email{mark.padgham@email.com}

}
\keyword{internal}
