% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apts.R
\name{cc_update_daily}
\alias{cc_update_daily}
\title{Update 'daily.md' file with new 'calcurse' entries}
\usage{
cc_update_daily(quiet = FALSE)
}
\arguments{
\item{quiet}{If 'TRUE', display information about file locations and statuses.}
}
\value{
(Invisibly) TRUE if 'daily.md' in the path specified by
'CALCURSE_DIR' is updated, otherwise FALSE.
}
\description{
Update 'daily.md' file with new 'calcurse' entries
}
