% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notes.R
\name{cc_update_notes}
\alias{cc_update_notes}
\title{Update 'daily.md' file from new 'calcurse' TODO notes}
\usage{
cc_update_notes(quiet = FALSE)
}
\arguments{
\item{quiet}{If 'TRUE', display information about file locations and statuses.}
}
\value{
(Invisibly) TRUE if 'daily.md' in the path specified by
'CALCURSE_DIR' is updated, otherwise FALSE.
}
\description{
Update 'daily.md' file from new 'calcurse' TODO notes
}
