% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcurseR-package.R
\docType{package}
\name{calcurseR-package}
\alias{calcurseR}
\alias{calcurseR-package}
\title{calcurseR: Functions to use 'calcurse' as a general organising tool}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Mark Padgham \email{mark.padgham@email.com}

}
\keyword{internal}
